\name{teminat-package}
\alias{teminat-package}
\alias{teminat}
\docType{package}
\title{
\packageTitle{teminat}
}
\description{
\packageDescription{teminat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{teminat}
\packageIndices{teminat}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{teminat}

Maintainer: \packageMaintainer{teminat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
